feat: mis reportes recientes

+++ src/main/java/com/itesm/domain/repository/ReportRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 public interface ReportRepository {
     Report save(Report report);
-    List<Report> findByUserId(Integer userId);
+    List<Report> findByUserId(Long userId);
 }
